refactor(pam): move getColumnIndicator to data-explorer-utils
