+++ frontend/src/pages/pam/PamDataExplorerPage/components/DataExplorerGrid.tsx
@@ -11,7 +11,7 @@ import { DataGrid, useDataGrid } from "@app/components/v3/generic/DataGrid";
 import { Skeleton } from "@app/components/v3/generic/Skeleton";
 
 import type { ColumnInfo, FieldInfo, ForeignKeyInfo, TableDetail } from "../data-explorer-types";
-import { getColumnIndicator } from "../data-explorer-types";
+import { getColumnIndicator } from "../data-explorer-utils";
 import type { FilterCondition, SortCondition } from "../sql-generation";
 import {
   buildCountQuery,
